Refactor UserInfo to user new UserDetails and UserUpdateRoles
UserInfo has been renamed UserUpdate
